Rename interactive-api-parent to interactive-api-lara-host
[#178367125]
