+++ lara-typescript/src/interactive-api-client/hooks.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { act, renderHook } from "@testing-library/react-hooks";
 import { render } from "@testing-library/react";
-import { mockIFramePhone, MockPhone } from "../interactive-api-parent/mock-iframe-phone";
+import { mockIFramePhone, MockPhone } from "../interactive-api-lara-host/mock-iframe-phone";
 import * as hooks from "./hooks";
 import * as iframePhone from "iframe-phone";
 import { getClient } from "./client";
