+++ lara-typescript/src/interactive-api-lara-host/global-iframe-saver.ts
