+++ lara-typescript/src/interactive-api-lara-host/mock-iframe-phone.ts
